+++ jaseci_core/jaseci/jac/ir/passes/__init__.py
@@ -1,0 +1,5 @@
+from jaseci.jac.ir.passes.ir_pass import IrPass  # noqa
+from jaseci.jac.ir.passes.prune_pass import PrunePass  # noqa
+from jaseci.jac.ir.passes.printer_pass import PrinterPass  # noqa
+from jaseci.jac.ir.passes.stats_pass import StatsPass  # noqa
+from jaseci.jac.ir.passes.codegen_pass import CodeGenPass  # noqa
