+++ jaseci_core/jaseci/jac/jsci_vm/tests/test_codegen.py
@@ -9,7 +9,7 @@ class TestCodegen(CoreTest):
 
     fixture_src = __file__
 
-    def test_walker_yield(self):
+    def test_simple_codegen(self):
         self.call(
             self.mast,
             ["sentinel_register", {"code": self.load_jac("simple.jac")}],
